[rename] 오타수정 Bycicle -> Bicycle
