+++ src/pages/PostPage/components/postForm/BicycleTypeInput/BicycleTypeInput.style.ts
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+`;
